Move get_display_axes into utils
